YaNav Updates
Removed YaNavBlockExtensions in favour of the functions being in
GyroExtensions, as will only use the rotate functions in conjunction
with existing functions in that file, and I see seldom use of the
existing functions without using the rotate functions.

YaNavControl:
Added Accuracy
